* clean DistantClient::processSend
> limit number of datagrams to send to 255 by using a uint8 parameter rather than size_t
>> should never be used with a high number anyway

+++ NetworkLib/src/UDP/DistantClient.cpp
@@ -138,7 +138,7 @@ namespace Bousk
 					// Error
 				}
 			}
-			void DistantClient::processSend(const size_t maxDatagrams /*= 0*/)
+			void DistantClient::processSend(const uint8 maxDatagrams /*= 0*/)
 			{
 				const auto now = Utils::Now();
 				// We do send data during connection process in order to keep it available before we accept it
